[Refactor/#51] rest uri naming conventions / application test / firebase 반영

+++ src/main/java/com/assu/server/domain/auth/security/jwt/JwtUtil.java
@@ -16,6 +16,7 @@ import io.jsonwebtoken.security.Keys;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -33,6 +34,7 @@ import java.util.concurrent.TimeUnit;
  */
 @Component
 @RequiredArgsConstructor
+@Profile("!test")
 public class JwtUtil {
 
     @Value("${jwt.secret}")
@@ -49,7 +51,9 @@ public class JwtUtil {
 
     @PostConstruct
     public void clearRedisOnStartup() {
-        redisTemplate.getConnectionFactory().getConnection().flushAll();
+        if (redisTemplate != null && redisTemplate.getConnectionFactory() != null) {
+            redisTemplate.getConnectionFactory().getConnection().flushAll();
+        }
     }
 
     // ───────── 토큰 생성 공통 유틸 ─────────
